Refactor to reflect new package name, moving load to GnocchiContext
