+++ gnocchi-core/src/main/scala/org/bdgenomics/gnocchi/association/LoadPhenotypesWithoutCovariates.scala
@@ -18,10 +18,10 @@
 package org.bdgenomics.gnocchi.association
 
 import htsjdk.samtools.ValidationStringency
-import net.fnothaft.gnocchi.models._
 import org.apache.spark.SparkContext
 import org.apache.spark.rdd.RDD
 import org.apache.spark.sql.{ Dataset, Row, SQLContext }
+import org.bdgenomics.gnocchi.models.Phenotype
 
 /*
 Takes in a text file containing phenotypes where the first line of the textfile is a header containing the phenotype lables.
