+++ gnocchi-core/src/main/scala/org/bdgenomics/gnocchi/association/SiteRegression.scala
@@ -17,7 +17,6 @@
  */
 package org.bdgenomics.gnocchi.association
 
-import net.fnothaft.gnocchi.models.{ Association, GenotypeState, Phenotype }
 import org.apache.spark.rdd.RDD
 import org.bdgenomics.adam.models.ReferenceRegion
 import org.bdgenomics.formats.avro.{ Contig, Variant }
