+++ gnocchi-core/src/main/scala/org/bdgenomics/gnocchi/transformations/Gs2variant.scala
@@ -17,7 +17,6 @@
  */
 package org.bdgenomics.gnocchi.transformations
 
-import net.fnothaft.gnocchi.models.GenotypeState
 import org.bdgenomics.formats.avro.{ Contig, Variant }
 
 object Gs2variant {
