+++ gnocchi-core/src/test/scala/org/bdgenomics/gnocchi/association/LoadPhenotypesWithoutCovariatesSuite.scala
@@ -17,12 +17,11 @@
  */
 package org.bdgenomics.gnocchi.association
 
-import net.fnothaft.gnocchi.GnocchiFunSuite
-import net.fnothaft.gnocchi.models.{ Phenotype }
 import org.apache.spark.sql.SQLContext
 import org.apache.spark.SparkContext
 import org.apache.spark.SparkContext._
 import org.apache.spark.SparkConf
+import org.bdgenomics.gnocchi.GnocchiFunSuite
 
 class LoadPhenotypesWithoutCovariatesSuite extends GnocchiFunSuite {
